+++ components/org.wso2.carbon.identity.api.user.push/org.wso2.carbon.identity.rest.api.user.push.v1/src/main/java/org/wso2/carbon/identity/rest/api/user/push/v1/impl/DiscoveryDataApiServiceImpl.java
@@ -18,12 +18,12 @@
 
 package org.wso2.carbon.identity.rest.api.user.push.v1.impl;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.wso2.carbon.identity.rest.api.user.push.v1.DiscoveryDataApiService;
 import org.wso2.carbon.identity.rest.api.user.push.v1.core.PushDeviceManagementService;
 
 import javax.ws.rs.core.Response;
 import org.wso2.carbon.identity.rest.api.user.push.v1.factories.PushDeviceManagementServiceFactory;
+import org.wso2.carbon.identity.rest.api.user.push.v1.model.DiscoveryDataDTO;
 
 /**
  * Implementation class of Push device Handler User APIs.
@@ -44,7 +44,7 @@ public class DiscoveryDataApiServiceImpl implements DiscoveryDataApiService {
     @Override
     public Response getRegistrationDiscoveryData() {
 
-        String responseDTO = pushDeviceManagementService.getRegistrationDiscoveryData();
-        return Response.ok().entity(responseDTO).build();
+        DiscoveryDataDTO discoveryDataDTO = pushDeviceManagementService.getRegistrationDiscoveryData();
+        return Response.ok().entity(discoveryDataDTO).build();
     }
 }
