+++ components/org.wso2.carbon.identity.api.user.push/pom.xml
@@ -35,10 +35,4 @@
         <module>org.wso2.carbon.identity.api.user.push.common</module>
     </modules>
 
-    <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
 </project>
